feat: evex-3 (15.74M / 147話者 / コーパス v4) を足す
ft-2 / ft-3 で測って効いたコーパスの工夫を、ゼロから学習する系にも持ち込む。
train 6,479,221 → 23,950,617 トークン (3.70倍)。

- 話者を 48 → 147 人 (発言の被覆 85.3% → 96.6%)
- 返信先を <|re|><|sM|> で残す。113,001 件中 94,729 件で相手が分かる
  (evex-2 は「返信かどうか」だけ持っていて相手を捨てていた)
- 窓を 60分/60件/3600字 に。切り方を 3 通りにして 2.08 倍
  ※ 開始位置をずらす方式は効かない (36,647 窓中 35,858 が同じ本文)。
    splitIntoChunks は主に沈黙時間で切るので境界が動かない
- 噛み合いと長い発言の切り出しを移植。判定は excerpts.mjs に共有化し、
  build-sft.mjs の出力が 1 バイトも変わらないことを md5 で確認した

推論側の噛み合わせも直した。どれも「例外は出ずに静かに崩れる」類:

- 文脈の話者の振り方が学習と違った。全員に役を配っていたので、名前を持つ人が
  学習時 <|s0|> / 推論時 <|a|> になっていた (147人ぶんが使われないまま)
- ownTurns が返信先を「次の話者」と読んで本文を丸ごと落としていた。
  そのせいで「記号だけの返答」が 42.2% に見えていた (直すと 0.0%)
- 本文中の <|...|> が制御記号に化けていた。実データに 16 件あり、
  train.txt の <|end|> が <|conv|> より 7 個多くなっていた
- server.py の話者走査が range(64) 固定で、147人だと 83 人が落ちていた

GPU まわり:

- T4 で bf16 を選んでいた。is_bf16_supported() は sm_75 でも True を返すが
  Turing の bf16 は tensor core に乗らない。世代で決めるようにして
  13,223 → 54,556 tok/s (4.1倍)
- epoch 終わりのサンプル生成が CPU テンソルを渡して落ちていた。
  チェックポイントは残るのに push まで進まず 38 分ぶんが消えた

+++ scripts/llm/model.py
@@ -70,9 +70,15 @@ def apply_rope(x, cos, sin):
     cos = cos[:t].view(1, 1, t, -1)
     sin = sin[:t].view(1, 1, t, -1)
 
-    even, odd = x[..., 0::2], x[..., 1::2]
+    # **回転は fp32 で計算して、最後に x の型に戻す。**
+    #
+    # cos/sin は fp32 の buffer なので、半精度の x と掛けると結果だけ fp32 に
+    # 昇格する。そのまま返すと q/k が fp32・v が半精度で
+    # scaled_dot_product_attention に入り、型が揃わない。
+    # 精度も落とさずに済むので、計算は fp32 のまま最後に揃える。
+    even, odd = x[..., 0::2].float(), x[..., 1::2].float()
     rotated = torch.stack((even * cos - odd * sin, even * sin + odd * cos), dim=-1)
-    return rotated.flatten(-2)
+    return rotated.flatten(-2).to(x.dtype)
 
 
 class Attention(nn.Module):
